moved isAuthorized in directive folder removed resolver index.ts and load it via glob pattern
